Same a messy fixed, but with additional F changes (untested)
Additional change
- for 'F' I changed the strides to 1 and kernels to 3 for a bigger output (untested)

Major Changes

the loss function is now hinge loss
fixed the output reshapes to keep batches
adjusted dimensions to fit fixes
Stuff that needs work in the near term:

As a consequence of fixing the output reshapes to maintain the batches, F and H are now outputting size 5. This means that J is EXTREMELY small - too small to probably learn anything (in the paper it's something like size 4096). This need to come up in size, so F and H need to be changed some to not reduce so much (F is the issue).
The dataset is very small. It should be able to memorize it, which should(?) be fine for testing. But, more images might be good. We could keep picking 'names' and loading images the way we do (in the ImageNet file) or we could iterate from the list - which would require some code rework but the parts are all there.