Pytorch_mini_BigBiGAN
This is a mini version of the BigBiGAN contain the base structure of the model:
- Encoder
- Generator
- F discriminator, simple CNN
- H discriminator, 3 layer FC
- J discriminators,  3 layer FC
- Sx, Sz and Sxz scores, FC down to a scalar

There is a error that I have not yet looked at that caused me to slot in BCE loss over the Loss_D and Loss_EG sections as a quick fix

The model generally trains, but some times goes unstable so that the generator mostly produces noise.
There is also a DCGAN in the model that can be used to compare results, since training can take a long time. If the results aren't inline with or exceeding the DCGAN by ~300 iterations, then it may not develop away from the noise.